pyproject.toml
[project]
name = "paper-supporter"
version = "1.0.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "openai (>=1.61.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyside6 (>=6.8.2,<7.0.0)",
    "markdown (>=3.7,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
